Reflecting the development results of NICT at FY2021-2022.

+++ bin/install.sh
@@ -3,15 +3,48 @@
 ORIGIN=`pwd`
 cd `dirname $0`
 
-curl -O http://download.redis.io/redis-stable.tar.gz
-tar xf redis-stable.tar.gz
-cd redis-stable
-make
-cp src/redis-server ../../redis/
-cd ..
-rm -rf redis-*
+privateip=127.0.0.1
 
-cd ..
+echo "Please input the private IP address for this host. (e.g. 192.168.xxx.xxx)"
+read privateip
+echo "Private IP : $privateip"
+
+# ssl
+while true;do
+    echo "Do you install a cert key for SSL? (yes or no)"
+    read answer
+    case $answer in
+        yes|y)
+			echo "creating cert.."
+			sh create_local_cert.sh $privateip
+            break
+            ;;
+        no|n)
+            break
+            ;;
+        *)
+            echo -e "cannot understand $answer.\n"
+            ;;
+    esac
+done
+
+# redis
+if [ "$(uname)" != "Darwin" ]; then
+    if [ ! -e "../redis/redis-server" ]; then
+        curl -O http://download.redis.io/redis-stable.tar.gz
+        tar xf redis-stable.tar.gz
+        cd redis-stable
+        make distclean
+        make
+        cp src/redis-server ../../redis/
+        cd ..
+        rm -rf redis-*
+    fi
+else
+    # mac
+    cp ../redis/redis-server.mac ../redis/redis-server
+fi
+
 npm install
 
 cd $ORIGIN
